requirements.txt
chromadb
pdfplumber
streamlit
ollama
python-docx
openpyxl
pandas
pypdf
